[Added]: Controller for root-mapping, property pipeline.is.component to the pipeline, [Deleted]: Current approach
