+++ qanary_component-template/src/main/java/eu/wdaqua/qanary/component/QanaryServiceController.java
@@ -1,56 +1,34 @@
 package eu.wdaqua.qanary.component;
 
-import java.net.URI;
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-
-import javax.inject.Inject;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
-
-import org.reflections.Reflections;
-import org.reflections.scanners.Scanners;
-import org.reflections.util.ConfigurationBuilder;
+import eu.wdaqua.qanary.commons.QanaryMessage;
+import eu.wdaqua.qanary.commons.QanaryUtils;
+import eu.wdaqua.qanary.commons.config.QanaryConfiguration;
+import io.swagger.v3.oas.annotations.Operation;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.core.env.Environment;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.ResponseBody;
 
-import eu.wdaqua.qanary.commons.QanaryMessage;
-import eu.wdaqua.qanary.commons.QanaryUtils;
-import eu.wdaqua.qanary.commons.config.QanaryConfiguration;
-import eu.wdaqua.qanary.component.exceptions.AmbiguousExtendingComponentClass;
-import eu.wdaqua.qanary.component.exceptions.NoExtendingComponentClass;
-import io.swagger.v3.oas.annotations.Operation;
-
+import javax.inject.Inject;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.net.URI;
 
 @Controller
 public class QanaryServiceController {
 
-    private static final Logger logger = LoggerFactory.getLogger(QanaryServiceController.class);
-    
     public final static String filenameOnlyPostInteractionAllowed = "only-post-is-allowed.html";
-
+    private static final Logger logger = LoggerFactory.getLogger(QanaryServiceController.class);
     @Value("${spring.boot.admin.client.url}")
     private String qanaryHost;
 
     private QanaryComponent qanaryComponent;
-    
-    @Autowired
-    private Environment env;
 
     @Inject
     public QanaryServiceController(QanaryComponent qanaryComponent) {
@@ -59,19 +37,18 @@ public class QanaryServiceController {
     }
 
     /**
-     * example: 
-     <pre>
-     	curl -X POST -d 'message={"http://qanary/#endpoint":"http://x.y"}' http://localhost:8080/annotatequestion | python -m json.tool
-     </pre>
-     *
+     * example:
+     * <pre>
+     * curl -X POST -d 'message={"http://qanary/#endpoint":"http://x.y"}' http://localhost:8080/annotatequestion | python -m json.tool
+     * </pre>
      */
-    @PostMapping(value = { QanaryConfiguration.annotatequestion, "/" + QanaryConfiguration.annotatequestion }, consumes = "application/json", produces = "application/json")
+    @PostMapping(value = {QanaryConfiguration.annotatequestion, "/" + QanaryConfiguration.annotatequestion}, consumes = "application/json", produces = "application/json")
     @ResponseBody
-	@Operation(
-			summary = "Each Qanary process will implement this endpoint as it is required ", //
-			operationId = "showDescriptionOnGetRequestOnRoot", //
-			description = "for showing information in a Web browser" //  
-	)
+    @Operation(
+            summary = "Each Qanary process will implement this endpoint as it is required ", //
+            operationId = "showDescriptionOnGetRequestOnRoot", //
+            description = "for showing information in a Web browser" //
+    )
     public ResponseEntity<?> annotatequestion(HttpServletRequest request, @RequestBody String message) throws Exception {
         logger.info("annotatequestion: {}", message);
         long start = QanaryUtils.getTime();
@@ -81,17 +58,17 @@ public class QanaryServiceController {
         QanaryConfiguration.setServiceUri(new URI(qanaryHost));
 
         QanaryMessage myQanaryMessage = new QanaryMessage(message);
-        
+
         this.qanaryComponent.setQanaryMessage(myQanaryMessage);
         this.qanaryComponent.setUtils(myQanaryMessage);
-        
+
         try {
             this.qanaryComponent.process(myQanaryMessage);
-		} catch (Exception e) {
-			logger.error("Something went wrong while executing the 'process' method: {}", e.getMessage());
-			return new ResponseEntity<String>(e.toString(), HttpStatus.INTERNAL_SERVER_ERROR);
-		}
-        
+        } catch (Exception e) {
+            logger.error("Something went wrong while executing the 'process' method: {}", e.getMessage());
+            return new ResponseEntity<String>(e.toString(), HttpStatus.INTERNAL_SERVER_ERROR);
+        }
+
 
         logger.debug("processing took: {} ms", QanaryUtils.getTime() - start);
 
@@ -100,100 +77,13 @@ public class QanaryServiceController {
 
     /**
      * not intended -> fallback: showing information to the user
-     * 
+     *
      * @param response
      * @return
      * @throws Exception
      */
-    @GetMapping(value =  { QanaryConfiguration.annotatequestion, "/" + QanaryConfiguration.annotatequestion } )
+    @GetMapping(value = {QanaryConfiguration.annotatequestion, "/" + QanaryConfiguration.annotatequestion})
     public String showDescriptionOnGetRequest(HttpServletResponse response) throws Exception {
-    	return filenameOnlyPostInteractionAllowed;
-    }
-
-    /**
-     * fallback: showing description
-     * 
-     * @param response
-     * @return
-     * @throws Exception
-     */
-    @GetMapping(value = "/")
-	@Operation(
-			summary = "Show description of component (HTML page)", //
-			operationId = "showDescriptionOnGetRequestOnRoot", //
-			description = "for showing information in a Web browser" //  
-	)
-    public String showDescriptionOnGetRequestOnRoot(HttpServletResponse response, Model model, HttpSession session) throws Exception {
-    	
-    	List<String> envImportantPropertyNames = Arrays.asList("spring.application.name", "spring.application.description", "springdoc.swagger-ui.path", "spring.boot.admin.url");
-    	
-    	// collect all important properties and add them to the session attributes
-    	for (String name : envImportantPropertyNames) {
-    		// replace characters that will not be accepted by Thymeleaf
-    		String sessionDataName = name.replace(".", "_").replace("-", "_");
-    		// add all values to the session including a default text if the parameter is not available
-    		String sessionDataValue = env.getProperty(name, "This text is shown as the property " + name + " is not defined (e.g., in application.properties).");
-	    	session.setAttribute(sessionDataName, sessionDataValue);
-    		logger.info("session | {} -> {}={}", name, sessionDataName, session.getAttribute(sessionDataName));
-	}
-
-    	Map<String,String> envImportantPropertyNameValue = new HashMap<>();
-    	envImportantPropertyNameValue.put("component_description_url", QanaryConfiguration.description);
-    	envImportantPropertyNameValue.put("component_description_file", QanaryConfiguration.description_file);
-    	envImportantPropertyNameValue.put("rdfcomponentdescription", QanaryConfiguration.rdfcomponentdescription);
-    	envImportantPropertyNameValue.put("ImplementationVersion", getClass().getPackage().getImplementationVersion());
-    	envImportantPropertyNameValue.put("ImplementationTitle", getClass().getPackage().getImplementationTitle());
-    	envImportantPropertyNameValue.put("ImplementationVendor", getClass().getPackage().getImplementationVendor());
-    	envImportantPropertyNameValue.put("Name", getClass().getPackage().getName());
-    	envImportantPropertyNameValue.put("SpecificationTitle", getClass().getPackage().getSpecificationTitle());
-    	envImportantPropertyNameValue.put("SpecificationVendor", getClass().getPackage().getSpecificationVendor());
-    	envImportantPropertyNameValue.put("SpecificationVersion", getClass().getPackage().getSpecificationVersion());
-
-    try {
-        //find the specific implemented Qanary component
-        Class<? extends QanaryComponent> extendingComponent = getExtendingComponent();
-        envImportantPropertyNameValue.put("componentImplementationVersion", extendingComponent.getPackage().getImplementationVersion());
-        envImportantPropertyNameValue.put("componentImplementationTitle", extendingComponent.getPackage().getImplementationTitle());
-        envImportantPropertyNameValue.put("componentImplementationVendor", extendingComponent.getPackage().getSpecificationVendor());
-
-    } catch (Exception e) {
-        logger.warn("No class implementing QanaryComponent could be found during runtime!");
-        logger.warn(e.getMessage());
-    }
-
-    	for (Map.Entry<String, String> entry : envImportantPropertyNameValue.entrySet()) {
-		String key = entry.getKey();
-		String val = entry.getValue();
-	    	session.setAttribute(key, val);
-    		logger.info("session | {}={}", key, session.getAttribute(key));
-	}
-    	
-    	return QanaryConfiguration.description_file;
-    }
-
-    /**
-     * get the specific implemented component that extends eu.wdaqua.qanary.component
-     * 
-     * @return
-     * @throws Exception
-     */
-    private Class<? extends QanaryComponent> getExtendingComponent() throws Exception {
-        Reflections reflections = new Reflections(
-                new ConfigurationBuilder()
-                    .addScanners(Scanners.SubTypes.filterResultsBy(s->true))
-                    .forPackages("eu.wdaqua.qanary.component")); 
-        // use reflections to find subtypes of eu.wdaqua.qanary.component
-        Set<Class<? extends QanaryComponent>> classes = reflections.getSubTypesOf(QanaryComponent.class);
-        // exactly one class is expected
-        if (classes.size() == 1) {
-            logger.debug("Found class: {}", classes.iterator().next().getName());
-            logger.debug("version: {}", classes.iterator().next().getPackage().getImplementationVersion());
-            // return the first (and only) one
-            return classes.iterator().next();
-        } else if (classes.size() == 0) {
-            throw new NoExtendingComponentClass(QanaryComponent.class);
-        } else {
-            throw new AmbiguousExtendingComponentClass(QanaryComponent.class);
-        }
+        return filenameOnlyPostInteractionAllowed;
     }
 }
