+++ qanary_pipeline-template/pom.xml
@@ -261,6 +261,12 @@
             <artifactId>qald.evaluator</artifactId>
             <version>[1.0.2,2.0.0)</version>
         </dependency>
+        <dependency>
+            <groupId>eu.wdaqua.qanary</groupId>
+            <artifactId>qa.component</artifactId>
+            <version>3.10.0</version>
+            <scope>compile</scope>
+        </dependency>
     </dependencies>
 
     <!-- START - Setup for deployment to maven repository -->
