--- qanary_pipeline-template/src/main/java/eu/wdaqua/qanary/QanaryServiceControllerConfiguration.java
@@ -1,20 +1,0 @@
-package eu.wdaqua.qanary;
-
-import eu.wdaqua.qanary.component.QanaryComponent;
-import eu.wdaqua.qanary.component.QanaryServiceController;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Primary;
-
-@Configuration
-public class QanaryServiceControllerConfiguration {
-
-    @Bean
-    @Primary
-    public QanaryServiceController qanaryServiceController(QanaryComponent qanaryComponent) {
-        return new QanaryServiceControllerWrapper(qanaryComponent);
-    }
-
-
-
-}
